adding 104-fibonacci.c file

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,20 +1,23 @@
 #include <stdio.h>
+#include "main.h"
 
 /**
 * main - prints the first 98 Fibonacci numbers
 * Return: Always 0
 */
 
+
+
 int main(void)
 {
 int i;
-long int first = 1, second = 2, next;
+long long int first = 1, second = 2, next;
 
-printf("%ld, %ld, ", first, second);
+printf("%lld, %lld, ", first, second);
 
 for (i = 3; i <= 98; i++) {
 next = first + second;
-printf("%ld", next);
+printf("%lld", next);
 
 if (i != 98)
 printf(", ");
